refactor: move v2 index to v2 package
