+++ src/main/kotlin/com/asyncapi/plugin/idea/extensions/index/v2/AsyncAPISpecificationIndex.kt
@@ -1,4 +1,4 @@
-package com.asyncapi.plugin.idea.extensions.index
+package com.asyncapi.plugin.idea.extensions.index.v2
 
 import com.intellij.json.JsonFileType
 import com.intellij.util.indexing.*
